Adds wrapper for panels to check authentication
